removed test functions from header file
