+++ src/wasm/tree_parse.cpp
@@ -1,36 +1,7 @@
 #include <iostream>
-#include <fstream>
-#include <nlohmann/json.hpp>
 #include <vector>
 #include <unordered_map>
 #include "tree_parse.hpp"
-
-using json = nlohmann::json;
-
-
-TreeNode parse_json_to_tree(const json& j_node) {
-    TreeNode node;
-    node.name = j_node.value("name", "");
-    node.value = 0;
-    
-    if (j_node.contains("value")) {
-        if (j_node["value"].is_number_unsigned()) {
-            node.value = j_node["value"].get<uint64_t>();
-        } else if (j_node["value"].is_number_integer()) {
-            node.value = static_cast<uint64_t>(j_node["value"].get<int64_t>());
-        }
-    }
-
-    node.cold = j_node.value("cold", false);
-    node.left_sum = 0;
-
-    if (j_node.contains("children")) {
-        for (const auto& child : j_node["children"]) {
-            node.children.push_back(parse_json_to_tree(child));
-        }
-    }
-    return node;
-}
 
 uint64_t calculate_left_sum(TreeNode& node, uint64_t& running_sum) {
     if (node.children.empty()) {
@@ -64,36 +35,4 @@ TreeNode prune_tree(const TreeNode& node, uint64_t threshold_left, uint64_t thre
     pruned_node.value = node.value - extra_time_left - extra_time_right; //substract extra left + right time intervals from walltime if any
 
     return pruned_node;
-}
-
-json tree_to_json(const TreeNode& node) {
-    json j_node;
-    j_node["name"] = node.name;
-    j_node["value"] = node.value;
-    j_node["left_sum"] = node.left_sum;
-    j_node["cold"] = node.cold;
-
-    for (const auto& child : node.children) {
-        j_node["children"].push_back(tree_to_json(child));
-    }
-
-    return j_node;
-}
-
-void save_json_to_file(const json& j, const std::string& filename) {
-    std::ofstream file(filename);
-    if (file.is_open()) {
-        file << j.dump();
-        file.close();
-    } else {
-        std::cerr << "Unable to open file: " << filename << std::endl;
-    }
-}
-
-void print_tree(const TreeNode& node, int level) {
-    std::string indent(level * 2, ' ');
-    std::cout << indent << "Name: " << node.name << ", Value: " << node.value << ", Left Sum: " << node.left_sum << "\n";
-    for (const auto& child : node.children) {
-        print_tree(child, level + 1);
-    }
 }
