+++ src/wasm/tree_parse.hpp
@@ -2,14 +2,8 @@
 #ifndef TREE_PARSE_H
 #define TREE_PARSE_H
 
-#include <iostream>
-#include <fstream>
-#include <nlohmann/json.hpp>
 #include <vector>
 #include <unordered_map>
-
-using json = nlohmann::json;
-
 
 struct TreeNode {
     std::string name;
@@ -19,16 +13,9 @@ struct TreeNode {
     std::vector<TreeNode> children;
 };
 
-TreeNode parse_json_to_tree(const json& j_node);
-
 uint64_t calculate_left_sum(TreeNode& node, uint64_t& running_sum);
 
 TreeNode prune_tree(const TreeNode& node, uint64_t threshold_left, uint64_t threshold_right);
 
-json tree_to_json(const TreeNode& node);
-
-void save_json_to_file(const json& j, const std::string& filename);
-
-void print_tree(const TreeNode& node, int level = 0);
 
 #endif // TREE_PARSE_H
